requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
marshmallow==3.13.0
python-dotenv==1.0.1
alembic==1.13.2
Werkzeug==3.0.3
